api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distlib==0.4.0
ecdsa==0.19.2
fastapi==0.135.1
fastapi-swagger-ui-theme==0.1.0
filelock==3.25.2
h11==0.16.0
idna==3.11
packaging==26.0
pipenv==2026.2.1
platformdirs==4.9.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-discovery==1.2.0
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
virtualenv==21.2.0
